producer/requirements.txt
aiohttp==3.8.1
aioresponses==0.7.3
asynctest==0.13.0
kafka-python==2.0.2
pytest==7.0.1
pytest-asyncio==0.16.0
